Film service without JUnit tests

+++ src/main/java/com/ula/filmstar/repository/FilmRepository.java
@@ -6,5 +6,5 @@ import com.ula.filmstar.model.Film;
 
 public interface FilmRepository extends JpaRepository<Film, Long> {
 
-	Film findById(int id);
+	Film findById(Long id);
 }
